requirements.txt
absl-py==2.0.0
aniso8601==9.0.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.3
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.2
exceptiongroup==1.1.3
Flask==2.3.3
flask-restx==1.2.0
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
gunicorn==21.2.0
h5py==3.10.0
idna==3.4
importlib-resources==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
keras==2.14.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyproject_hooks==1.0.0
pyrsistent==0.20.0
pytest==7.4.3
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
rsa==4.9
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==2.3.7
wrapt==1.14.1
